Retrain Inception Model
Added a new page which allows the user to retrain the inception model with their own sets of images. This is still in beta and needs finishing. The screen can be used only once python and pip are installed. The environment path variables for pip and python need to be set manually before using the screen.