Added Close button and search bar
I have created a custom tabitem which along with a close button functionality. A search bar has also been added to the main window where the user can search for pages in the software or the internet. The other pages have been optimised for performance. Lastly, the search bar is experimental and will need to be optimised in the future.